+++ routes/init.rb
@@ -1,4 +1,8 @@
 # encoding: utf-8
 require_relative 'customers'
 require_relative 'accounts'
+require_relative 'tasks'
+require_relative 'jobs'
+require_relative 'settings'
+require_relative 'wordlists'
 #require_relative 'main'
