+++ routes/main.rb
@@ -1,0 +1,4 @@
+# encoding: utf-8
+class HashView < Sinatra::Application
+
+end
